% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltt_lineup_plots.R
\name{lineup_keyplot}
\alias{lineup_keyplot}
\title{Make lineup plot solution key}
\usage{
lineup_keyplot(lineup_dat)
}
\arguments{
\item{lineup_dat}{dataframe structured from \code{\link{make_lineup_dat}}}
}
\value{
ggplot
}
\description{
Function to return list of plot types related to order of plots, used in assessment step
}
\examples{
dat=data.frame(x=rnorm(50))
dat$y=x+rnorm(50)
lineup_dat = make_lineup_dat(M=8, dat, xname="x", yname="y")
lineup_keyplot(lineup_dat)
}
