% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltt_pvals.R
\name{ltt}
\alias{ltt}
\title{Lady tasting tea p-values}
\usage{
ltt(M, x)
}
\arguments{
\item{M}{is the number of plot pairs in a line-up}

\item{x}{the number of incorrect choices}
}
\value{
lady tasting tea p-value
}
\description{
Function for looking up ltt probabilities
}
