% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lineup_data.R
\name{make_alt_dat}
\alias{make_alt_dat}
\title{Function to make alternate plot data}
\usage{
make_alt_dat(dat, xname = "x", yname = "y")
}
\arguments{
\item{dat}{dataframe}

\item{xname}{column name from dat for x-dimension in plots (defaults "x")}

\item{yname}{column name from dat for y-dimension in plots (defaults "y")}
}
\description{
capture mean structure between X and Y with local linear
 regression with optimal bandwidth, then permute residuals
}
\examples{
dat=data.frame(x=rnorm(5))
dat$y=x+rnorm(5)
alt_dat <- make_alt_dat(dat)
}
