% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lineup_data.R
\name{make_lineup_dat}
\alias{make_lineup_dat}
\title{Make a lineup plot data}
\usage{
make_lineup_dat(M, dat, xname = "x", yname = "y", seed = NULL)
}
\arguments{
\item{M}{number of pairs of plots}

\item{dat}{dataframe}

\item{xname}{column name from dat for x-dimension in plots (defaults "x")}

\item{yname}{column name from dat for y-dimension in plots (defaults "y")}

\item{seed}{set random seed for reproducible lineups}
}
\value{
dataframe with columns x, permy, type, order
}
\description{
Function to create stacked dataframe with M pairs of plots
 based on original dataframe named dat. ItWill randomize the
 order labels used in building plot lineup
}
\examples{
dat=data.frame(x=rnorm(5))
dat$y=x+rnorm(5)
lineup_dat = make_lineup_dat(M=8, dat, xname="x", yname="y")
}
