% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltt_pvals.R
\name{sim_curve}
\alias{sim_curve}
\title{Generate data from quadratic curve}
\usage{
sim_curve(n, b0, b1, b2, sd = 1, range = c(-3, 3))
}
\arguments{
\item{n}{sample size}

\item{b0}{intercept of curve}

\item{b1}{first order poly term}

\item{b2}{second order poly term}

\item{sd}{standard deviation for random noise around curve}

\item{range}{range of x values over which to generate data}
}
\value{
dataframe of simulated x,y pairs
}
\description{
Function to simulate from quadratic curve
}
\examples{
curve_dat <- sim_curve(n=100, b0=0, b1=0, b2=.3, sd=2.6, range=c(-3,3))
plot(curve_dat)
summary(lm(y ~ I(x^2), data=curve_dat))
}
