% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltt_pvals.R
\name{tea_pval_table}
\alias{tea_pval_table}
\title{Organize lady tasting tea p-values into TeX table}
\usage{
tea_pval_table(M_range = 5:10, x_range = 0:3)
}
\arguments{
\item{M_range}{numeric vector with the range of possible plot pairs in a line-up}

\item{x_range}{numeric vector with the range for number of incorrect choices}
}
\value{
TeX table
}
\description{
Organize lady tasting tea p-values into TeX table
}
\examples{
tea_pval_table(5:10,0:2)
}
