% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltt_lineup_plots.R
\name{tt_lineup_plot}
\alias{tt_lineup_plot}
\title{Function to generate tea-tasting lineup}
\usage{
tt_lineup_plot(lineup_dat)
}
\arguments{
\item{lineup_dat}{dataframe structured from \code{\link{make_lineup_dat}}}
}
\value{
ggplot
}
\description{
Function to generate tea-tasting lineup
}
\examples{
dat=data.frame(x=rnorm(50))
dat$y=x+rnorm(50)
lineup_dat = make_lineup_dat(M=8, dat, xname="x", yname="y")
tt_lineup_plot(lineup_dat)
}
