% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lineup_data.R
\name{make_lineup_dat}
\alias{make_lineup_dat}
\title{Make a lineup plot data}
\usage{
make_lineup_dat(M, dat, xname = "x", yname = "y", seed = NULL,
  bootstrap = FALSE, jitter = 0, both = FALSE)
}
\arguments{
\item{M}{number of pairs of plots}

\item{dat}{dataframe}

\item{xname}{column name from dat for x-dimension in plots (defaults "x")}

\item{yname}{column name from dat for y-dimension in plots (defaults "y")}

\item{seed}{set random seed for reproducible lineups}

\item{bootstrap}{logical indicator of if (bootstrap==TRUE) bootstrapping used, if (bootstrap==FALSE) then LOESS residuals permuted}

\item{jitter}{amount of jitter to add to point locations - units of sds}

\item{both}{logical indicator to run both LOESS residual permutation and bootstrapping of points}
}
\value{
dataframe with columns x, permy, type, order
}
\description{
Function to create stacked dataframe with M pairs of plots
 based on original dataframe named dat. ItWill randomize the
 order labels used in building plot lineup
}
\examples{
dat=data.frame(x=rnorm(25))
dat$y=dat$x+rnorm(25)
lineup_dat = make_lineup_dat(M=8, dat, xname="x", yname="y")
lineup_dat = make_lineup_dat(M=8, dat, xname="x", yname="y",bootstrap=T,jitter=T)
}
